solved a few compiler warnings (ncurses uses exclusively int type and not size_t, etc)

+++ src/cross.c
@@ -36,12 +36,12 @@ void cross(int x, int y)
     #define CROSS_CENTRE_STR "X"
 
     //get maximums from screen size (it updates the vars)
-   size_t max_y = 0;
-    size_t max_x = 0;
+    int max_y = 0;
+    int max_x = 0;
     getmaxyx(stdscr, max_y, max_x);
 
     //VERTICAL
-    for(size_t i = 0; i < max_y; i++)
+    for(size_t i = 0; i < (size_t)max_y; i++)
     {
         int num_to_print = abs((int)-i + (max_y/2));
